requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via django
asttokens==2.4.1
    # via stack-data
billiard==4.2.0
    # via celery
celery[redis]==5.4.0
    # via
    #   -r requirements.in
    #   celery
    #   django-celery-beat
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.3
    # via django-celery-beat
decorator==5.1.1
    # via ipython
django==5.0.7
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-picklefield
    #   django-timezone-field
    #   djangorestframework
django-celery-beat==2.6.0
    # via -r requirements.in
django-constance==3.1.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-picklefield==3.2
    # via django-constance
django-timezone-field==7.0
    # via django-celery-beat
djangorestframework==3.15.2
    # via -r requirements.in
executing==2.0.1
    # via stack-data
gunicorn==22.0.0
    # via -r requirements.in
ipython==8.26.0
    # via -r requirements.in
jedi==0.19.1
    # via ipython
kombu==5.3.7
    # via celery
matplotlib-inline==0.1.7
    # via ipython
packaging==24.1
    # via gunicorn
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
prompt-toolkit==3.0.47
    # via
    #   click-repl
    #   ipython
psycopg[binary,pool]==3.2.1
    # via
    #   -r requirements.in
    #   psycopg
psycopg-binary==3.2.1
    # via psycopg
psycopg-pool==3.2.2
    # via psycopg
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.18.0
    # via ipython
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   python-crontab
redis==5.0.7
    # via celery
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sqlparse==0.5.1
    # via django
stack-data==0.6.3
    # via ipython
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   ipython
    #   psycopg
    #   psycopg-pool
tzdata==2024.1
    # via
    #   celery
    #   django-celery-beat
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise==6.7.0
    # via -r requirements.in
